backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
python-dotenv==0.17.1
Flask-JWT-Extended==4.3.1
Flask-Bcrypt==0.7.1
Flask-CORS==3.0.10